feat: implement session management components, payment status flow, a…

+++ src/components/layout/Initializer.tsx
@@ -1,19 +1,25 @@
 "use client";
 import { INACTIVE_SESSION_STATUS, useSession } from "@/contexts";
 import { useRouter } from "next/navigation";
-import { useEffect } from "react"
+import { useEffect, useRef } from "react"
 import toast, { Toaster } from 'react-hot-toast';
 
 
 function Initializer() {
     const { clearSession, sessionData } = useSession();
+    const ref = useRef<boolean>(false);
     const router = useRouter();
     useEffect(() => {
         if (sessionData?.session?.status && INACTIVE_SESSION_STATUS.includes(sessionData?.session?.status as any)) {
-            clearSession();
-            localStorage.clear();
-            toast.success('Session ended!', { duration: 5000, position: "top-center" });
-            router.push("/");
+            if (!ref.current) {
+                ref.current = true;
+                setTimeout(() => {
+                    clearSession();
+                    localStorage.clear();
+                    toast.success('Session ended!', { duration: 5000, position: "top-center" });
+                    router.push("/");
+                }, 5000);
+            }
         }
     }, [sessionData?.session?.status]);
     return (
